% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMSfunctions.R
\name{AdjustedCurieWeissfit}
\alias{AdjustedCurieWeissfit}
\title{Modified Curie-Weiss fitting function}
\usage{
AdjustedCurieWeissfit(
  Tstart = 0,
  Tend = 300,
  material = "",
  dataframe,
  Tcolumn = "TemperatureK",
  Susccolumn = "susc",
  ChiInvcolumn = "chiinv",
  chi0guess = 0,
  Cconstguess = 0.8,
  thetaCWguess = 1000,
  color = "darkgray",
  dofit = FALSE,
  pubreadyplot = FALSE
)
}
\arguments{
\item{Tstart}{The temperature which should be the lower bound of the Curie-Weiss analysis}

\item{Tend}{The temperature which should be the upper bound of the Curie-Weiss analysis}

\item{material}{A string describing the material to go in plot titles}

\item{dataframe}{Name of dataframe containing columns of data. Note that if your data contains the same temperatures repeated, you should use dataframe[x:y,] to select the points x-y to use.}

\item{Tcolumn}{String name of column in dataframe containing temperature data}

\item{Susccolumn}{String name of column in dataframe containing susceptibility data}

\item{chi0guess}{Numeric initial guess of the chi0 value. Linear data should have chi0=0. Note that a wildly off guess can cause this function to fail.}

\item{Cconstguess}{Numeric initial guess of the Curie-Weiss parameter value (in units determined by your susceptibility data). Note that a wildly off guess can cause this function to fail.}

\item{thetaCWguess}{Numeric initial guess of the Curie-Weiss temperature value (in units of your temperature data). This is the y-intercept of your data. Note that a wildly off guess can cause this function to fail.}

\item{color}{Color to plot the data in.}

\item{dofit}{Boolean for whether to perform a nonlinear fit. If 'FALSE', a linear fit only is attempted to provide R^2 values (see ?BasicCurieWeissfit)}

\item{pubreadyplot}{Boolean for whether to open a popup window with fit parameters in a text box on top of the plot in the style favored by McQueen Lab for publications.}

\item{ChiInvColumn}{String name of column in dataframe containing inverse susceptibility data (used to perform a traditional Curie-Weiss fit)}
}
\description{
This function performs a nonlinear fit to magnetic susceptibility data and plots it along with various fit parameters. See https://doi.org/10.1038/s42005-022-00853-y
}
