% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC functions.R
\name{ApplyW2Approx}
\alias{ApplyW2Approx}
\title{x^2 low-frequency approximator}
\usage{
ApplyW2Approx(EstEnd, stepsize = 1, origdata, xtitle = "x")
}
\arguments{
\item{EstEnd}{The point where you want the parabolic function to end (does not have to be a real value in the data, the function which match the closest point)}

\item{stepsize}{step size of points between your end pt and the origin}

\item{origdata}{a dataframe containing your original data}

\item{xtitle}{a string matching the column name for the independent variable in your dataframe}
}
\value{
Dataframe containing a column "x" that contains the sequence of x-axis values and a colummn "y" that contains the y-axis values.
}
\description{
This function takes in (x,y) data and replaces all x data before some threshold value EstEnd with a parabolic curve connecting 0 to the original function's value at that point.
}
\examples{
x <- seq(0,100,0.5)
y <- x^2*sin(x)
data <- data.frame(x,y)
newdata <- ApplyW2Approx(45,1,data,"x")
plot(newdata$x,newdata$y) # see ?plot
lines(data$x,data$y,col=2) # see ?lines
}
