% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMSfunctions.R
\name{BasicCurieWeissfit}
\alias{BasicCurieWeissfit}
\title{Curie-Weiss fitting function}
\usage{
BasicCurieWeissfit(
  Tstart = 0,
  Tend = 300,
  material = "",
  dataframe,
  Tcolumn = "TemperatureK",
  ChiInvcolumn = "chiinv",
  color = "darkgray",
  pubreadyplot = FALSE
)
}
\arguments{
\item{Tstart}{The temperature which should be the lower bound of the Curie-Weiss analysis}

\item{Tend}{The temperature which should be the upper bound of the Curie-Weiss analysis}

\item{material}{A string describing the material to go in plot titles}

\item{dataframe}{Name of dataframe containing columns of data. Note that if your data contains the same temperatures repeated, you should use dataframe[x:y,] to select the points x-y to use.}

\item{Tcolumn}{String name of column in dataframe containing temperature data}

\item{color}{Color to plot the data in.}

\item{pubreadyplot}{Boolean for whether to open a popup window with fit parameters in a text box on top of the plot in the style favored by McQueen Lab for publications.}

\item{ChiInvColumn}{String name of column in dataframe containing inverse susceptibility data}
}
\description{
This function performs a linear fit to magnetic susceptibility data and plots it along with various fit parameters.
}
