% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMSfunctions.R
\name{CalculateSusceptibility}
\alias{CalculateSusceptibility}
\title{Magnetic susceptibility calculator}
\usage{
CalculateSusceptibility(
  dataframe,
  H = "MagneticFieldOe",
  M = "DCMomentFreeCtremu",
  MolarMass = 1,
  Mass = 1,
  CalcInv = TRUE
)
}
\arguments{
\item{dataframe}{Name of dataframe containing columns of data}

\item{H}{String name of column in dataframe containing applied magnetic field data (take note of your units; this function does no conversion)}

\item{M}{String name of column in dataframe containing sample magnetization data (take note of your units; this function does no conversion)}

\item{MolarMass}{Numeric molar mass of sample (assumed in amu)}

\item{Mass}{Numeric mass of sample (assumed in grams)}

\item{CalcInv}{By default, the function also returns 1/susceptibility unless this parameter is set to 'FALSE'}
}
\value{
Input dataframe with magnetic susceptibility ("Susc") and, optionally, inverse susceptibility ("Chiinv") columns attached at end
}
\description{
This function reads columns of data for applied field strength and magnetization and calculates the molar magnetic susceptibility
}
