% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMSfunctions.R
\name{ChiNaughtGuesser}
\alias{ChiNaughtGuesser}
\title{Modified Curie-Weiss analysis Chi-Naught guesser}
\usage{
ChiNaughtGuesser(
  Tstart = 0,
  Tend = 300,
  dataframe,
  Tcolumn = "TemperatureK",
  Susccolumn = "susc",
  initialguess = 0,
  guessforme = TRUE,
  limitto = "POS",
  perturb = 1.1
)
}
\arguments{
\item{Tstart}{The temperature which should be the lower bound of the Curie-Weiss analysis}

\item{Tend}{The temperature which should be the upper bound of the Curie-Weiss analysis}

\item{dataframe}{Name of dataframe containing columns of data. Note that if your data contains the same temperatures repeated, you should use dataframe[x:y,] to select the points x-y to use.}

\item{Tcolumn}{String name of column in dataframe containing temperature data}

\item{Susccolumn}{String name of column in dataframe containing susceptibility data}

\item{initialguess}{Numeric initial guess of the chi0 value. Linear data should have chi0=0. Note that a wildly off guess can cause this function to fail.}

\item{guessforme}{Boolean for whether to do the binary search for an improved chi0 value. If 'FALSE', this function returns the R^2 value of data and whether it "thinks" your chi0 guess can be improved.}

\item{limitto}{Options are "POS" for positive chi0 (concave data) and "NEG" for negative chi0 (convex data)}

\item{perturb}{Numeric value for the width of the window to search with the binary search algorithm. A typical value on first run is 1.5, resulting in a search window of -1.5*guess to 1.5*guess. After you get close to a good chi0, reduce the factor to 1.05, 1.005, or similar to maximize the number of sig figs in your answer.}
}
\value{
The numeric value of the best chi0 acheived by the function. Printed text comments guide the user in using the function to refine this value.
}
\description{
This function uses a binary search algorithm to guess the chi0 value required to make "almost linear" magnetic susceptibility data linear and thus enable Curie-Weiss analysis. See https://doi.org/10.1038/s42005-022-00853-y for a justification of this technique.
}
