% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGA functions.R
\name{ConvertToMassPercent}
\alias{ConvertToMassPercent}
\title{Convert Mass to Mass Percent}
\usage{
ConvertToMassPercent(
  dataframe,
  column = "Weight (mg)",
  outputname = "Relative Weight (\%)"
)
}
\arguments{
\item{dataframe}{Dataframe containing mass data.}

\item{column}{String name of column in the dataframe containing the mass data.}

\item{outputname}{String name of new column to create for the mass percent data.}
}
\value{
Dataframe containing all its original columns plus new column appended to the end.
}
\description{
This function calculates mass percent given a column of mass data. The first data point is assumed to be 100%.
}
