% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cubicity metric.R
\name{CubicDeviationMetric}
\alias{CubicDeviationMetric}
\title{Cubic Deviation Metric calculator (newest)}
\usage{
CubicDeviationMetric(a, b, c, al, be, ga, decimals = 4)
}
\arguments{
\item{a}{a lattice parameter, typically in angstroms}

\item{b}{b lattice parameter, typically in angstroms}

\item{c}{c lattice parameter, typically in angstroms}

\item{al}{alpha lattice parameter (angle between b and c sides) in degrees}

\item{be}{beta lattice parameter (angle between a and c sides) in degrees}

\item{ga}{gamma lattice parameter (angle between a and b sides) in degrees}

\item{decimals}{Number of decimal places to round answer to. Default is 4.}
}
\value{
Value of cubicity metric
}
\description{
This function calculates the value of the cubicity metric described in xxx.
}
\details{
The returned value of the cubicity metric will be zero if and only if the lattice parameters input correspond to a cubic unit cell. The returned value will be unitless and normalized such that it can be compared to values for other unit cells, with a possible maximum of 1. The closer to zero, the "more cubic" your unit cell; the closer to one, the "less cubic". It is important to note that multiple unit cells may produce the same metric value.
}
\examples{
library(tidyverse)
CubicDeviationMetric(10,10,10,90,90,90,decimals=3)
CubicDeviationMetric(10,8,9,63,70,63,decimals=3)
df <- data.frame(a=c(10,10),b=c(10,8),c=c(10,9),al=c(90,63),be=c(90,70),ga=c(90,63))
df <- df \%>\% mutate(CDM=CubicDeviationMetric(a,b,c,al,be,ga,4))
head(df)
}
