% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cubicity metric.R
\name{CubicityMetric}
\alias{CubicityMetric}
\title{Cubic Deviation Metric calculator (scalar; old)}
\usage{
CubicityMetric(a, b, c, al, be, ga, sigfigs = 4)
}
\arguments{
\item{a}{a lattice parameter, typically in angstroms}

\item{b}{b lattice parameter, typically in angstroms}

\item{c}{c lattice parameter, typically in angstroms}

\item{al}{alpha lattice parameter (angle between b and c sides) in degrees}

\item{be}{beta lattice parameter (angle between a and c sides) in degrees}

\item{ga}{gamma lattice parameter (angle between a and b sides) in degrees}

\item{sigfigs}{Number of sig figs to round answer to. Default is 4.}
}
\value{
Value of cubicity metric
}
\description{
This function calculates the value of the cubicity metric described in http://dx.doi.org/10.1038/s41535-025-00782-3. This function is deprecated but preserved for the purposes of interpreting the data in that paper.
}
\details{
The returned value of the cubicity metric will be zero if and only if the lattice parameters input correspond to a cubic unit cell. The returned value will be unitless and normalized such that it can be compared to values for other unit cells. The closer to zero, the "more cubic" your unit cell. It is important to note that multiple unit cells may produce the same metric value.
}
