% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic data functions.R
\name{curve.area}
\alias{curve.area}
\title{Numeric integrator (constant return)}
\usage{
curve.area(x, y, const.step = FALSE)
}
\arguments{
\item{x}{list of independent values}

\item{y}{list of dependent values}

\item{const.step}{Boolean descriptor for whether (x,y) pairs are evenly spaced. If you do not know, set this to FALSE. In the case the pairs are evenly spaced, the function will use a slightly faster integration method, but the standard method works will work in both cases.}
}
\value{
Integral (aka area under the curve) of input data.
}
\description{
This function performs a numeric integration. In other words, it calculates area under curve defined by series of (x,y) points. The method used is a trapezoidal Riemann sum.
}
\examples{
x<-seq(1,10,1)
y<-x^2
curve.area(x,y,const.step=TRUE)
curve.area(x,y,const.step=FALSE)
}
