% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC functions.R
\name{Cv.Tlist}
\alias{Cv.Tlist}
\title{Cv vs. T calculator}
\usage{
Cv.Tlist(f, D, Tmin = 2, Tmax = 300, Tstep = 1)
}
\arguments{
\item{f}{list of frequencies; typically input as column in dataframe. Should be in units of Hz or s^-1. *Please note* that this list is assumed to have a constant spacing between points.}

\item{D}{list of intensities; typically input as column in dataframe. Should be in units of s.}

\item{Tmin}{lowest temperature to calculate in Kelvin}

\item{Tmax}{highest temperature to calculate in Kelvin}

\item{Tstep}{stepsize between each temperature to calculate. For example, with Tmin=1, Tmax=5, Tstep=1... the function will return Cv's for T=1,2,3,4,5}
}
\value{
Dataframe containing column for Temperature in units of K, column for Constant volume heat capacity (Cv) in units of J/K, and column for Constant volume heat capacity divided by temperature cubed (CvperT3) in units of J/K^4
}
\description{
This function calculates constant volume heat capacity from phonon density of states as a function of frequency (in Hz) for a range of temperatures. It also calculates Cv/T^3. Please see http://lampz.tugraz.at/~hadley/ss1/dbr/dos2cv.html for functional form.
}
