% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC functions.R
\name{GaussbyFWHM}
\alias{GaussbyFWHM}
\title{Gaussian function in terms of FWHM}
\usage{
GaussbyFWHM(x, center, FWHM)
}
\arguments{
\item{x}{x-axis values}

\item{center}{Center position (in units of your x-axis) for the Gaussian}

\item{FWHM}{The full width at half maximum value for your Gaussian in units of your x-axis}
}
\value{
The value of a Gaussian defined by input parameters at the input x value.
}
\description{
This function represents a normalized Gaussian function (integral = 1) in terms of its center point and full width at half maximum (FWHM), instead of the traditional representation by standard deviation.
}
\examples{
GaussbyFWHM(1,0,10)

x <- seq(-20,20,0.5)
y <- GaussbyFWHM(x,0,10)
plot(x,y)
}
