% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGA functions.R
\name{LoadTGAData}
\alias{LoadTGAData}
\title{TGA data loader - binary}
\usage{
LoadTGAData(filename, densedata = TRUE, masspercent = TRUE)
}
\arguments{
\item{filename}{String name of the file to open. If the file is not stored in your current working directory, be sure to include the full path to the file.}

\item{densedata}{Boolean for whether to load all the data ('TRUE') or 1/5 data points ('FALSE'). The TA Univeral Analysis program only loads 1/5 data points from the original binary file by default, so to reproduce a .txt file from this program (or to reduce the memory used by R), you should set this to 'FALSE'.}

\item{masspercent}{Boolean for whether to convert mass data into weight percent data. If 'TRUE', both columns are saved, slightly increasing memory usage.}
}
\value{
Dataframe containing all columns in the datafile, plus optionally a weight pecent column.
}
\description{
This function loads into memory the data contained in a binary file generated by a TA SDT Q600 TGA-DSC. These files typically end in .001, .002, or similar. It may be applicable to other binary files with a header followed by numeric data, but this is untested. The "\f" character is used to detect the end of the header and start of the data.
}
