% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGA functions.R
\name{LoadTGADataTxt}
\alias{LoadTGADataTxt}
\title{TGA data loader - text files}
\usage{
LoadTGADataTxt(filename, masspercent = TRUE)
}
\arguments{
\item{filename}{String name of the file to open. If the file is not stored in your current working directory, be sure to include the full path to the file.}

\item{masspercent}{Boolean for whether to convert mass data into weight percent data. If 'TRUE', both columns are saved, slightly increasing memory usage.}
}
\value{
Dataframe containing all columns in the datafile, plus optionally a weight pecent column.
}
\description{
This function loads into memory the data contained in a text file generated by TA Universal Analysis based on binary file from a TA SDT Q600 TGA-DSC.
}
\details{
Exported TA SDT Q600 TGA-DSC files end in .txt but their contents may vary depending on what the Universal Analysis user selected during export. The "StartOfData" phrase is used to detect the end of the header and start of the data.
}
