% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGA functions.R
\name{LoadTGAHeader}
\alias{LoadTGAHeader}
\title{TGA file header loader - binary}
\usage{
LoadTGAHeader(filename)
}
\arguments{
\item{filename}{String name of the file to open. If the file is not stored in your current working directory, be sure to include the full path to the file.}
}
\value{
String vector containing all the header info.
}
\description{
This function loads into memory the header of a binary file generated by a TA SDT Q600 TGA-DSC. These files typically end in .001, .002, or similar. It may be applicable to other binary files with a header followed by numeric data, but this is untested. The "\f" character is used to detect the end of the header and start of the data.
}
