% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGA functions.R
\name{LoadTGAHeaderTxt}
\alias{LoadTGAHeaderTxt}
\title{TGA file header loader - text files}
\usage{
LoadTGAHeaderTxt(filename)
}
\arguments{
\item{filename}{String name of the file to open. If the file is not stored in your current working directory, be sure to include the full path to the file.}
}
\value{
String vector containing all the header info.
}
\description{
This function loads into memory the header of a text file generated by TA Universal Analysis based on binary file from a TA SDT Q600 TGA-DSC.
}
\details{
Exported TA SDT Q600 TGA-DSC files end in .txt but their contents may vary depending on what the Universal Analysis user selected during export. The "StartOfData" phrase is used to detect the end of the header and start of the data.
}
