% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC functions.R
\name{ManualSmear}
\alias{ManualSmear}
\title{Spectrum smearing function}
\usage{
ManualSmear(plotmin, plotmax, plotstep, FWHM, data)
}
\arguments{
\item{plotmin}{minimum value to calculate}

\item{plotmax}{maximum value to calculate}

\item{plotstep}{step size between minimum and maximum to calculate}

\item{FWHM}{The full width at half maximum value for your Gaussians in units of your x-axis}

\item{data}{list of x-axis values, typically a column in a dataframe}
}
\value{
Dataframe containing a column "x" that contains the sequence of x-axis values and a colummn "y" that contains the y-axis values computed from this sequence.
}
\description{
This function takes peak positions represented as equally-weighted lines and smears them out to a phonon or photon spectrum closer to that expected from experiment. This is accomplished by applying Gaussian functions to "smear out" each line. A la https://mdommett.github.io/blog/interpolation-with-gaussian-broadening/
}
\details{
Note that the peaks are all assumed to have equal weight. For an unequally-weighted function, see ?ManualSmearWeights.
}
\examples{
x <- c(30,45,67,88,89,90)
y <- c(1,1,1,1,1,1)
data <- ManualSmear(0,100,1,5,x)
plot(data)
points(x,y,col=2,pch=25)
}
