% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC functions.R
\name{ManualSmearWeights}
\alias{ManualSmearWeights}
\title{Spectrum smearing function (incl. peak weights)}
\usage{
ManualSmearWeights(plotmin, plotmax, plotstep, FWHM, data, weights)
}
\arguments{
\item{plotmin}{minimum value to calculate}

\item{plotmax}{maximum value to calculate}

\item{plotstep}{step size between minimum and maximum to calculate}

\item{FWHM}{The full width at half maximum value for your Gaussian in units of your x-axis}

\item{data}{list of x-axis values, typically a column in a dataframe}

\item{weights}{list of weightings (y-axis values) for each x-axis value, typically a column in a dataframe. Must have the same length as the x value list.}
}
\value{
Dataframe containing a column "x" that contains the sequence of x-axis values and a colummn "y" that contains the y-axis values computed from this sequence.
}
\description{
This function takes peak positions represented as lines with weights and smears them out to a phonon or photon spectrum closer to that expected from experiment. A la https://mdommett.github.io/blog/interpolation-with-gaussian-broadening/
}
\examples{
x <- c(30,45,67,88,89,90)
y <- c(3,2,1,4,1,6)
data <- ManualSmearWeights(0,100,1,5,x,y)
plot(data)
points(x,y,col=2,pch=25)
}
