% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic data functions.R
\name{numeric.derivative}
\alias{numeric.derivative}
\title{Numeric differentiator (dataset return)}
\usage{
numeric.derivative(x, y)
}
\arguments{
\item{x}{list of independent values}

\item{y}{list of dependent values}
}
\value{
Integral of input data as a vector of new derivative values. First point in list will be NA.
}
\description{
This function performs a numeric differentiation. In other words, it calculates the numeric derivative of the input data.
}
