% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMSfunctions.R
\name{OpenMPMSData}
\alias{OpenMPMSData}
\title{MPMS data loader}
\usage{
OpenMPMSData(
  folder,
  filename,
  removenames = c("default"),
  preserveempty = FALSE
)
}
\arguments{
\item{folder}{String describing folder containing the datafile}

\item{filename}{String of datafile name, should end in .dat}

\item{removenames}{Vector of strings to filter out of column names, i.e. words to search for in column names; these columns will not be loaded. Example: removenames = c("Comment","AC") will not load the Comment column nor the columns for AC Moment (emu), AC M. Std Err. (emu), AC Phase (deg), etc. etc.}

\item{preerveempty}{Boolean for whether to preserve columns which contain NO data.}
}
\value{
Dataframe containing selected columns in the .dat file
}
\description{
This function loads in data from a .dat file generated by a Quantum Design MPMS 3
}
