% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPMS functions.R
\name{OpenPPMSData}
\alias{OpenPPMSData}
\title{PPMS data loader}
\usage{
OpenPPMSData(
  folder = getwd(),
  filename = NULL,
  removenames = c("Drive", "Fit", "Center", "Count", "Gain", "Bridge", "Stat", "Meas",
    "Pressure", "Min", "Max", "Phase", "Elap", "Map", "Stamp", "PPMS"),
  preserveempty = FALSE,
  relativetime = FALSE
)
}
\arguments{
\item{folder}{String describing folder containing the datafile}

\item{filename}{String of datafile name, should end in .dat}

\item{removenames}{Vector of strings to filter out of column names, i.e. words to search for in column names; these columns will not be loaded. Example: removenames = c("Frequency","Amp") will not load the Frequency column nor the Amplitude column. The default is to remove columns describing instrument status and columns used to calculate observables (with the exception of the calculated straw center position). To **not** remove any columns, input an empty vector c().}

\item{relativetime}{Boolean for whether to rescale the time stamp to begin at 0 and count from there, creating a new column called 'AbsTimeSec'. Default is to instead report the PPMS measure of absolute time. To remove the old column of absolute time, add "Stamp" or similar to the vector in 'removenames'.}

\item{preerveempty}{Boolean for whether to preserve columns which contain NO data.}
}
\value{
Dataframe containing selected columns in the .dat file.
}
\description{
This function loads in data from a .dat file generated by a Quantum Design PPMS.
}
