% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC functions.R
\name{PhononDOStoCv}
\alias{PhononDOStoCv}
\title{Cv Calculator}
\usage{
PhononDOStoCv(f, D, T)
}
\arguments{
\item{f}{list of frequencies; typically input as column in dataframe. Should be in units of Hz or s^-1. *Please note* that this list is assumed to have a constant spacing between points.}

\item{D}{list of intensities; typically input as column in dataframe. Should be in units of s.}

\item{T}{single temperature in K}
}
\value{
Constant volume heat capacity (Cv) in units of J/K
}
\description{
This function calculates constant volume heat capacity from phonon density of states as a function of frequency (in Hz) for a single temperature. Please see http://lampz.tugraz.at/~hadley/ss1/dbr/dos2cv.html for functional form.
}
