% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMSfunctions.R
\name{ReadMPMSSampleMass}
\alias{ReadMPMSSampleMass}
\title{MPMS sample mass parser}
\usage{
ReadMPMSSampleMass(folder, filename)
}
\arguments{
\item{folder}{String describing folder containing the datafile}

\item{filename}{String of datafile name, should end in .dat}
}
\value{
Numeric sample mass in grams (not milligrams)
}
\description{
This function reads the header from a .dat file generated by a Quantum Design MPMS 3 and returns the user-input sample mass
}
