% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic data functions.R
\name{Riemann.sum}
\alias{Riemann.sum}
\title{Numeric integrator (dataset return)}
\usage{
Riemann.sum(x, y)
}
\arguments{
\item{x}{list of independent values}

\item{y}{list of dependent values}
}
\value{
Integral of input data as a vector of new integral values. First point in list will be NA.
}
\description{
This function performs a numeric integration (quadrature) of a series of (x,y) points. The method used is a trapezoidal Riemann sum.
}
