% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGA functions.R
\name{TGA2AxisPlot}
\alias{TGA2AxisPlot}
\title{Two-axis plot maker}
\usage{
TGA2AxisPlot(
  dataframe,
  title = "Placeholder title",
  colorpalette = "Dark2",
  x = "Time (min)",
  y = list("Temperature (°C)", "Weight (mg)"),
  width = 4,
  height = 3
)
}
\arguments{
\item{dataframe}{Dataframe containing data.}

\item{title}{String to be used as plot title}

\item{colorpalette}{Palette of colors to be used for the first and second lines in the plot. See ?palette for examples. Your own palette can be defined using c("color1","color2")}

\item{x}{String name of column in the dataframe containing the independent variable.}

\item{y}{Vector of strings containing the names of columns to be used as the 2 dependent variables.
#' @param width width of output image}

\item{height}{height of output image}
}
\description{
This function generates a plot with two scatterplots of data sharing an independent variable. These plots are commonly made by TA Universal Analysis but may be more generally applicable.
}
