% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMSfunctions.R
\name{CalculateSusceptibility}
\alias{CalculateSusceptibility}
\title{DC [Molar/Mass] Magnetic susceptibility calculator}
\usage{
CalculateSusceptibility(
  dataframe,
  H = "MagneticFieldOe",
  M = "DCMomentFreeCtremu",
  MolarMass = 1,
  Mass = 1,
  CalcInv = TRUE
)
}
\arguments{
\item{dataframe}{Name of dataframe containing columns of data}

\item{H}{String name of column in dataframe containing applied magnetic field data (take note of your units; this function does no conversion)}

\item{M}{String name of column in dataframe containing sample magnetization data (take note of your units; this function does no conversion)}

\item{MolarMass}{Numeric molar mass of sample (assumed in amu)}

\item{Mass}{Numeric mass of sample (assumed in grams)}

\item{CalcInv}{By default, the function also returns 1/susceptibility unless this parameter is set to 'FALSE'}
}
\value{
Input dataframe with magnetic susceptibility ("Susc") and, optionally, inverse susceptibility ("Chiinv") columns attached at end
}
\description{
This function reads columns of data for applied field strength and magnetization and calculates the standard, molar, or mass magnetic susceptibility assuming M=Chi*H (linear relationship valid for low fields in DC magnetization measurements).
}
\details{
If data were input in consistent units - either all cgs (emu for magnetization and Oe for field H) or all SI (A/m or J/Tm^3 for magnetization and A/m for field H) - then the units output are detailed below, and depend on which type of magnetic susceptibility you have chosen to calculate - standard, mass susceptibility, or molar susceptibility.

In the event you want susceptibility, not molar/mass susceptibility, set both 'Mass' and 'MolarMass' to 1. The resulting output value will then be unitless (if using cgs units, the output may also be represented as emu/cm^3, which is actually still unitless).

If you would like mass susceptibility, set 'MolarMass' to 1. The output units of this function for mass susceptibility are m^3/g (SI) and emu/g (cgs). 1 emu/g (cgs) = 4*pi*10^-3 m^3/kg (SI)

If you would like molar susceptibility, set neither 'Mass' nor 'MolarMass' to 1. The output units of this function for molar susceptibility are m^3/mol (SI) and emu/mol (cgs). 1 emu/mol (cgs) = 4*pi*10^-6 m^3/mol (SI)
}
