% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPMSfunctions.R
\name{OpenMPMSData}
\alias{OpenMPMSData}
\title{MPMS data loader}
\usage{
OpenMPMSData(
  folder = getwd(),
  filename = NULL,
  removenames = c("Action", "Range", "Min", "Max", "Pressure", "Number", "code", "Motor",
    "Stat", "Average", "Scan", "Chamber", "Drift", "Position", "Fit", "Phase", "ACX",
    "Averaging", "Cycles", "Type", "Trim", "Signal"),
  preserveempty = FALSE,
  relativetime = FALSE
)
}
\arguments{
\item{folder}{String describing folder containing the datafile}

\item{filename}{String of datafile name, should end in .dat}

\item{removenames}{Vector of strings to filter out of column names, i.e. words to search for in column names; these columns will not be loaded. Example: removenames = c("Comment","AC") will not load the Comment column nor the columns for AC Moment (emu), AC M. Std Err. (emu), AC Phase (deg), etc. etc. The default is to remove columns describing instrument status and columns used to calculate observables (with the exception of the calculated straw center position). To **not** remove any columns, input an empty vector c().}

\item{relativetime}{Boolean for whether to rescale the time stamp to begin at 0 and count from there, creating a new column called 'AbsTimeSec'. Default is to instead report the MPMS measure of absolute time. To remove the old column of absolute time, add "Stamp" or similar to the vector in 'removenames'.}

\item{preerveempty}{Boolean for whether to preserve columns which contain NO data.}
}
\value{
Dataframe containing selected columns in the .dat file.
}
\description{
This function loads in data from a .dat file generated by a Quantum Design MPMS 3.
}
\details{
The function will also print a warning message if a quick check of the DC Moment goodness of fit indicates a potential problem with the voltage fitting. It is good practice to review this data independently to confirm the validity of your moment results. See https://qdusa.com/siteDocs/appNotes/1500-018.pdf for more info.
}
