% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cubicity metric.R
\name{UnitCellVolume}
\alias{UnitCellVolume}
\title{Unit cell volume calculator}
\usage{
UnitCellVolume(a, b, c, al, be, ga, sigfigs = 4)
}
\arguments{
\item{a}{a lattice parameter, typically in angstroms}

\item{b}{b lattice parameter, typically in angstroms}

\item{c}{c lattice parameter, typically in angstroms}

\item{al}{alpha lattice parameter (angle between b and c sides) in degrees}

\item{be}{beta lattice parameter (angle between a and c sides) in degrees}

\item{ga}{gamma lattice parameter (angle between a and b sides) in degrees}

\item{sigfigs}{Number of sig figs to round answer to. Default is 4.}
}
\value{
Volume of the unit cell in (units of a/b/c)^3
}
\description{
This function calculates the volume of the unit cell
}
